First commit to GitHub
Previous implementations were done directly on the Raspberry Pi Zero W and tested on-site.
Before that, an Arduino Leonardo was used, alongside a DS1307 RTC (wouldn't recommend it) and a 2-relay module.
The code was ported to Python, the module was upgraded to a 4-relay version, and no RTC is used. Time is fetched automatically via internet.